#14 Made types work, added comments to the code, got rid of unnecessary code and simplified the code base
