+++ contract/src/lendingPool/priceManagerContract.js
@@ -7,16 +7,14 @@ import { makeScalarMap } from '@agoric/store';
 import { E } from '@agoric/eventual-send';
 
 /**
- * This is a faucet that provides liquidity for the ertp asset created
- * using the parameter in terms. Just for demonstration purposes.
+ * This the place where we wrap the PriceManager inside a vat.
+ * @see {PriceManager} to learn more about the functionality.
  */
 
 /** @type {ContractStartFn} */
 export async function start(zcf) {
 
   const priceAuthorities = makeScalarMap('priceAuthorities');
-  /** @type {MapStore<string, InnerLoan>} */
-  const supportedAssetPublicFacets = makeScalarMap('supportedAssetPublicFacets');
 
   /**
    * @param {Brand} brandIn
@@ -33,23 +31,15 @@ export async function start(zcf) {
     return notifier;
   }
 
-  const addNewSupportedAssetPublicFacet = (key, value) => {
-    supportedAssetPublicFacets.init(key, value);
-  }
-
-  const getPriceAuthority = (key) => {
+  const getWrappedPriceAuthority = (key) => {
     return priceAuthorities.get(key);
   }
 
-  const getsupportedAssetPublicFacets = (key) => {
-    return supportedAssetPublicFacets.get(key);
-  }
 
+  /** @type ERef<PriceManager>*/
   const publicFacet = Far('PriceManager', {
     addNewWrappedPriceAuthority,
-    addNewSupportedAssetPublicFacet,
-    getPriceAuthority,
-    getsupportedAssetPublicFacets
+    getWrappedPriceAuthority,
   })
 
   const creatorFacet = Far('faucetInvitationMaker', {});
